refactored core.reflection TypeInfo added new string tools for converting wchar to char strings
git-svn-id: https://subversion.assembla.com/svn/dsmks/branches/core_refac@673 c65b9dfa-2e9b-4d30-9b60-b5bbc3021b56

+++ src/scripting.python/PythonVariable.cpp
@@ -14,7 +14,7 @@ PythonVariable::PythonVariable(const std::string  &name){
   // get namespace / module string
   if(nameParts.size()>1){
     stringstream ss;
-    for(int i=0; i < nameParts.size()-1;i++){
+    for(size_t i=0; i < nameParts.size()-1;i++){
       ss << nameParts[i];
       ss << ".";
     }
